Sem2 added

+++ Seminar 1/Task4.py
@@ -1,7 +1,7 @@
 n = int(input('Введите число N: '))
 print('Введите сравниваемые числа: ')
-t = int(input())
+t = float(input())
 for i in range(1, n):
-    inp = int(input())
+    inp = float(input())
     t = inp if (inp > t) else t
 print('Наибольшее число: ', t)
